src/requirements.txt
Flask==3.0.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
gunicorn==23.*
